♻️Refactor: 영화검색 코드 Util로 분리, SearchMovie 중복코드 삭제
